--- a/packages/babel-traverse/src/path/context.js
+++ b/packages/babel-traverse/src/path/context.js
@@ -215,7 +215,6 @@ export function pushContext(context) {
 }
 
 export function setup(parentPath, container, listKey, key) {
-  this.inList = !!listKey;
   this.listKey = listKey;
   this.parentKey = listKey || key;
   this.container = container;