<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/master/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Minor: Bug Fix? | 👍
| Major: Breaking Change?  | No if one is using public API.
| Tests Added + Pass?      | Yes
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->

In this PR we improve the performance of babel-traverse by revising the memory layout of `NodePath`. It results to 10% performance gain compared to the control branch on the following test cases.

```js
const babel = require("/path/to/local/babel-core");
const code = require("fs").readFileSync("./fixtures/es5/babylon-dist.js");
babel.transform(code, { preset: ["@babel/env"] });
```

I can observe this branch outperforms control branch consistently.
```
┌─────────────────────┬──────────────────────────────┐
│ fixture             │ devTransform                 │
├─────────────────────┼──────────────────────────────┤
│ es5/babylon-dist.js │ 1.02 ops/sec ±31.84% (978ms) │
└─────────────────────┴──────────────────────────────┘
┌─────────────────────┬───────────────────────────────┐
│ fixture             │ baseTransform                 │
├─────────────────────┼───────────────────────────────┤
│ es5/babylon-dist.js │ 0.86 ops/sec ±20.52% (1167ms) │
└─────────────────────┴───────────────────────────────┘

```

And the memory usage improvements due to compressing traverse flags and adding accessor methods for `inList` and `parentPath`.
```
┌─────────────────────┬──────────────────┐
│ fixture             │ devTransform x5  │
├─────────────────────┼──────────────────┤
│ es5/babylon-dist.js │ heap: 239.37 MiB │
└─────────────────────┴──────────────────┘
┌─────────────────────┬──────────────────┐
│ fixture             │ baseTransform x5 │
├─────────────────────┼──────────────────┤
│ es5/babylon-dist.js │ heap: 305.47 MiB │
└─────────────────────┴──────────────────┘
```

The following changes are implemented:
- Removed `NodePath.inList` data property, a get accessor is provided for backward compatibility (partially). 
- Removed `NodePath.parentKey` data property, a get accessor is provided for backward compatibility (partially). 
- Removed `NodePath.typeAnnotation` data property, it is introduced in #1752 but since then it is not used by any routines. I believe it is a mistype of `Node.typeAnnotation`.
- Compressed `NodePath.shouldSkip/shouldStop/removed` into `_traverseFlags` bit array, both get accessor and set accessor is provided for backward compatibility.
- Lazily assign `this.skipKeys` in `setContext`.
- Lazily initialize `this.data` in `NodePath.constructor`.

In practice `inList` and `parentKey` should be treated as read-only since these two properties are rewritten in `setContext`. The current behavior accepts one assigning these two properties.

The `shouldSkip/shouldStop/removed` traverse flags are handy and turbofan would inline these simple accessors we don't introduce performance overhead here. By compressing these states into bit array we optimize the performance of `setContext`.