@JLHwung 2019-09-22
@nicolo-ribaudo I am sorry I make a typo before. It should be the `parentKey` instead of `parentPath`, but your point is still valid on `parentKey`.

> I think that any plugin relying on the old behavior was actually broken: if someone wanted to change the parent node by doing path.parentKey = ... it wouldn't have worked.

I agree with you that setting `inList` and/or `parentKey` never works in practice. Instead of failing silently, babel will now report `can not assign to an accessor`. In this way it seems more like a bug fix, though. 😁